Explicit cast to int

+++ src/rdkafka_range_assignor.c
@@ -79,7 +79,7 @@ rd_kafka_range_assignor_assign_cb (rd_kafka_t *rk,
                  * each consumer. */
                 numPartitionsPerConsumer =
                         eligible_topic->metadata->partition_cnt /
-                        rd_list_cnt(&eligible_topic->members);
+                        (int) rd_list_cnt(&eligible_topic->members);
 
                 /* If it does not evenly divide, then the first few consumers
                  * will have one extra partition. */
